sonarLint issues fixed & small refactorings
