+++ echo-service/src/main/java/org/exampledriven/docker/echo/EchoServiceApplication.java
@@ -1,0 +1,19 @@
+package org.exampledriven.docker.echo;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+
+@EnableWebMvc
+@SpringBootApplication
+@EnableDiscoveryClient
+public class EchoServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(EchoServiceApplication.class, args);
+    }
+
+}
+
+
