+++ echo-service/src/main/java/org/exampledriven/docker/echo/domain/HostFactory.java
@@ -10,6 +10,9 @@ import javax.servlet.http.HttpServletRequest;
  */
 public class HostFactory {
 
+    private HostFactory(){
+    }
+
     public static HostInfo create(HttpServletRequest request) {
         return new HostInfo(request.getLocalAddr(), request.getLocalPort(), getHostName());
     }
