--- eureka-server/src/main/java/org/exampledriven/eureka/Application.java
@@ -1,17 +1,0 @@
-package org.exampledriven.eureka;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
-
-@SpringBootApplication
-@EnableEurekaServer
-@EnableDiscoveryClient
-public class Application {
-
-	public static void main(String[] args) throws Exception {
-		SpringApplication.run(Application.class, args);
-	}
-
-}
