+++ eureka-server/src/main/java/org/exampledriven/eureka/EurekaServerApplication.java
@@ -1,20 +1,17 @@
-package org.exampledriven.docker.echo;
+package org.exampledriven.eureka;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
 
-@EnableWebMvc
 @SpringBootApplication
+@EnableEurekaServer
 @EnableDiscoveryClient
-public class Application {
+public class EurekaServerApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
+        SpringApplication.run(EurekaServerApplication.class, args);
     }
 
 }
-
-
